Split the admin questions template test into focused suites
The single file was ~862 lines, over the ~400-line target. Each page now has
its own suite under test/ui/templates/admin/questions/, sharing one fixtures
helper. Assertions are unchanged.
