+++ test/ui/templates/admin/questions/fixtures.ts
@@ -1,0 +1,21 @@
+import {
+  smallLargeAnswers,
+  testListingWithCount,
+  testQuestion,
+} from "#test-utils/factories.ts";
+import { setupFeaturePageTest } from "../feature-page-test.ts";
+
+export const TEST_LISTINGS = [
+  testListingWithCount({ id: 1, name: "Spring Gig" }),
+  testListingWithCount({ id: 2, name: "Summer Gig" }),
+];
+
+/** The "T-shirt size?" question with Small/Large answers — the canonical radio
+ *  question reused by the question, answer-edit, and answer-delete page tests. */
+export const tShirtQuestion = testQuestion({
+  answers: smallLargeAnswers,
+  id: 1,
+  text: "T-shirt size?",
+});
+
+export const setupQuestionPageTest = setupFeaturePageTest("questions");
